// web/bs-web/src/app/models/BookView.ts
export default class BookView{
    code: number;
    title: string;
    publisher: string;
    edition: number;
    year: number;
    authorCode: number;
    authorName: string;
    subjectCode: number;
    subjectDescription: string;
    priceTableCode: number;
    priceTableDescription: string;
    price: number;
}